Worked on the sailing files i.e
File I/O_sailings.h and .cpp -> i added some functions to the .h file then i did the implementation in the .cpp file for them.

Then i worked on the sailing.cpp file using the user manual as a reference for this. i changed the createsailing variables from a const and then also the sailingID to a string in the commontypes.h file.